celery_worker/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
amqp==5.3.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
frozenlist==1.5.0
geoip2==4.8.0
greenlet==3.1.1
idna==3.10
kombu==5.4.2
maxminddb==2.6.2
multidict==6.1.0
prompt_toolkit==3.0.48
propcache==0.2.0
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
redis==5.2.0
requests==2.32.3
sendgrid==6.11.0
setuptools==75.5.0
six==1.16.0
SQLAlchemy==2.0.36
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
tzdata==2024.2
ua-parser==0.18.0
urllib3==2.2.3
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.17.1
